Partially implemented character editing
